More performance tests to find issues with current c/s.
For single connection programs, the current c/s appears to be ok when compared with a couple different approaches to sending the objects. The results are not too much different, generally a bit better for querying, but current c/s is better for sets, which is most likely due to the new protocol reinstantiating objects on the server (some low level calls could probably fix this discrepancy). See ClientServerPerformanceTest.java.

Also tried a concurrency test and this is where the performance really degrades. The new protocols appear to be much faster, in the order of 300x for sets and 2x for query/delete. This shows that there are huge improvements to be had just in the client-server communications independent of core concurrency performance. see ProducerConsumerTest.java.